requirements.txt
termcolor==2.4.0
PyYAML==6.0.1
requests==2.31.0
beautifulsoup4==4.12.3
langchain_community
fastapi==0.111.0
ipython==8.12.3
langchain==0.2.5
langchain_experimental==0.0.61
langchain_groq==0.1.5
langchain_openai==0.1.8
pandas==2.2.2
pydantic==2.7.4
SQLAlchemy==2.0.30
streamlit==1.28.2
uvicorn==0.30.1